trip-builder-service/requirements.txt
isodate==0.6.0
geopy~=2.2.0
scipy~=1.7.2
Flask~=1.1.4
markupsafe~=2.0.1

python-dotenv~=0.19.0
requests~=2.26.0
pyfunctional==1.4.3

boto3~=1.21.6

PyYAML~=3.12

pytest
elasticsearch~=7.15.2

# py_zipkin==1.2.0
